feat: add game service player repository
